Move constants to changeme file
