+++ src/routes/(apps)/recalplus/+page.svelte
@@ -2,7 +2,7 @@
 import Calendar from "$lib/components/recal/Calendar.svelte";
 import Left from "$lib/components/recal/Left.svelte";
 import Top from "$lib/components/recal/Top.svelte";
-import { CURRENT_TERM_ID } from "$lib/constants";
+import { CURRENT_TERM_ID } from "$lib/changeme.js";
 import { isMobile, showCal } from "$lib/stores/mobile";
 import { rawCourseData, ready, schedules, searchCourseData } from "$lib/stores/recal.js";
 import { rMeta } from "$lib/stores/rmeta.js";
